\name{BMUHepta}
\alias{BMUHepta}

\docType{data}
\title{Best matching units (BMU) of Hepta from FCPS (Fundamental Clustering Problem Suite)}

\description{Best matching units (BMU) of an ESOM projection of the Hepta data set 
from FCPS (Fundamental Clustering Problem Suite) on an 80 x 40 planar grid of artifical neurons.}

\usage{data("BMUHepta")}

\details{
Size 212, Dimensions 3 (key, linecoordinates, columncoorditaes)

Classes 7, stored in \code{Hepta$Cls}
}

\references{
Ultsch A, Lotsch J: Machine-learned cluster identification in high-dimensional data. J Biomed Inform. 2017 Feb;66:95-104. doi: 10.1016/j.jbi.2016.12.011. Epub 2016 Dec 28. }
\examples{
data("BMUHepta")
str("BMUHepta")
}
\keyword{datasets Hepta FCPS ESOM}
