\name{Hepta}
\alias{Hepta}

\docType{data}
\title{Hepta from FCPS (Fundamental Clustering Problem Suite)}

\description{Dataset with 7 easily seperable classes.}

\usage{data("Hepta")}

\details{
Size 212, Dimensions 3, stored in \code{Hepta$Data}

Classes 7, stored in \code{Hepta$Cls}
}

\references{
Ultsch, A.: U* C: Self-organized Clustering with Emergent Feature Maps, Lernen, Wissensentdeckung und Adaptivitaet (LWA), pp. 240-244, Saarbruecken, Germany, 2005. }
\examples{
data(Hepta)
str(Hepta)
}
\keyword{datasets Hepta FCPS}
